Includes today's interval training.
My stats are pretty poor, although I did start my first interval on
the asphalt crossing in front of Albuquerque Academy.  That's the
furthest downhill that I'll start.  I always start on a multiple of
fifteen seconds (elapsed time) and had the chance to start right at 0
seconds, so I took it.

I *think* my watch can be programmed to give beeps for intervals, but
it's been a long time since I've looked into it and I figure I'm going
to get a new watch "soon" (but haven't for years), so I never spend
the time looking.