Includes today's (brutal) interval pack run.
It's nice to start doing these again, but I pretty much screwed up on
multiple fronts, including not thinking about my inability to get to
my iPod (since it's in a zippered compartment behind my head) to
switch from random music to speed 3, so I had to come to a complete
stop to make that adjustment.  Running out of water was not fun
either.

Plenty of room for improvement though!