Includes today's interval training.
To be honest, it was a little intimidating doing interval training
today on a sore foot with compromised lungs.  My intervals themselves
were very good considering, although my overall time for 15.66 miles
was 2:44:26, a 10:29 min/mile pace and that was at an average HR of
133.  I actually walked the final 1.38 miles home after doing the
intervals because my HR average was already 136bpm and I also wanted
to baby my foot as much as possible.